Rename project package naming from com.treelogic.proteus to eu.proteus.producer
