--- src/main/java/com/treelogic/proteus/exceptions/IllegalNumberOfColumnsException.java
@@ -1,10 +1,0 @@
-package com.treelogic.proteus.exceptions;
-
-public class IllegalNumberOfColumnsException extends Exception {
-
-	private static final long serialVersionUID = -1022961137880815777L;
-
-	public IllegalNumberOfColumnsException(String message) {
-		super(message);
-	}
-}
