--- src/main/java/com/treelogic/proteus/model/HSMMeasurementMapper.java
@@ -1,19 +1,0 @@
-package com.treelogic.proteus.model;
-
-public class HSMMeasurementMapper {
-
-	private static final String SPLITTER = (String) ProteusData.get("model.hsm.splitter");
-
-	public static HSMMeasurement map(String record) {
-		String[] lineSplit = record.split(SPLITTER);
-		int coilID = Integer.parseInt(lineSplit[0]);
-
-		HSMMeasurement hsmRecord = new HSMMeasurement(coilID);
-
-		for (String line : lineSplit) {
-			hsmRecord.put(line);
-		}
-
-		return hsmRecord;
-	}
-}
