--- src/main/java/com/treelogic/proteus/model/Measurement.java
@@ -1,5 +1,0 @@
-package com.treelogic.proteus.model;
-
-public abstract class Measurement {
-
-}
