+++ src/main/java/eu/proteus/producer/exceptions/IllegalNumberOfColumnsException.java
@@ -1,0 +1,10 @@
+package eu.proteus.producer.exceptions;
+
+public class IllegalNumberOfColumnsException extends Exception {
+
+	private static final long serialVersionUID = -1022961137880815777L;
+
+	public IllegalNumberOfColumnsException(String message) {
+		super(message);
+	}
+}
