+++ src/main/java/eu/proteus/producer/model/HSMMeasurementMapper.java
@@ -1,0 +1,19 @@
+package eu.proteus.producer.model;
+
+public class HSMMeasurementMapper {
+
+	private static final String SPLITTER = (String) ProteusData.get("model.hsm.splitter");
+
+	public static HSMMeasurement map(String record) {
+		String[] lineSplit = record.split(SPLITTER);
+		int coilID = Integer.parseInt(lineSplit[0]);
+
+		HSMMeasurement hsmRecord = new HSMMeasurement(coilID);
+
+		for (String line : lineSplit) {
+			hsmRecord.put(line);
+		}
+
+		return hsmRecord;
+	}
+}
