+++ src/main/java/eu/proteus/producer/model/Measurement.java
@@ -1,0 +1,5 @@
+package eu.proteus.producer.model;
+
+public abstract class Measurement {
+
+}
