+++ src/main/java/eu/proteus/producer/model/SensorMeasurement1D.java
@@ -1,0 +1,18 @@
+package eu.proteus.producer.model;
+
+public class SensorMeasurement1D extends SensorMeasurement {
+
+
+	public SensorMeasurement1D(){}
+	
+    public SensorMeasurement1D(int coilId, double x, int variableIdentifier, double value){
+        super();
+        this.coilId = coilId;
+        this.x = x;
+        this.varName = variableIdentifier;
+        this.value = value;
+    }
+   
+
+}
+
