+++ src/main/java/eu/proteus/producer/source/InMemorySourceStrategy.java
@@ -1,0 +1,5 @@
+package eu.proteus.producer.source;
+
+public class InMemorySourceStrategy {
+
+}
