+++ src/main/java/eu/proteus/producer/source/SourceStrategy.java
@@ -1,0 +1,14 @@
+package eu.proteus.producer.source;
+
+import java.util.stream.Stream;
+/**
+ * 
+ * @author nach0
+ *
+ * @param <T>
+ */
+public interface SourceStrategy <T>{
+
+	
+	public Stream<T> readFile(String filePath);
+}
