+++ src/main/java/eu/proteus/producer/tasks/ProteusFlatnessTask.java
@@ -1,0 +1,28 @@
+package eu.proteus.producer.tasks;
+
+import eu.proteus.producer.kafka.ProteusKafkaProducer;
+import eu.proteus.producer.model.SensorMeasurement;
+
+import java.util.List;
+
+
+public class ProteusFlatnessTask extends ProteusTask{
+	
+    /**
+     * List of "coiltimeseries" records containing flatness values
+     */
+    private List<SensorMeasurement> flatnessRows;
+
+    public ProteusFlatnessTask(List<SensorMeasurement> flatnessRows) {
+        super();
+        this.flatnessRows = flatnessRows;
+    }
+
+    @Override
+    public Void call() throws Exception {
+        this.flatnessRows
+                .stream()
+                .forEach(ProteusKafkaProducer::produceFlatness);
+        return null;
+    }
+}
